Moved mixins to proper location
