+++ src/main/java/net/crioch/fifymcc/mixin/screen/CraftingScreenHandlerMixin.java
@@ -1,4 +1,4 @@
-package net.crioch.fifymcc.mixin.screenhandler;
+package net.crioch.fifymcc.mixin.screen;
 
 import net.crioch.fifymcc.recipe.ComponentRecipeMatcher;
 import net.crioch.fifymcc.interfaces.ComponentRecipeInputProvider;
@@ -7,28 +7,28 @@ import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.inventory.RecipeInputInventory;
 import net.minecraft.recipe.Recipe;
 import net.minecraft.recipe.RecipeEntry;
-import net.minecraft.screen.PlayerScreenHandler;
+import net.minecraft.screen.CraftingScreenHandler;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 
-@Mixin(PlayerScreenHandler.class)
-public class PlayerScreenHandlerMixin implements ComponentRecipeScreenHandler<RecipeInputInventory> {
+@Mixin(CraftingScreenHandler.class)
+public class CraftingScreenHandlerMixin implements ComponentRecipeScreenHandler<RecipeInputInventory> {
     @Final
     @Shadow
-    private RecipeInputInventory craftingInput;
+    private RecipeInputInventory input;
 
     @Final
     @Shadow
-    private PlayerEntity owner;
+    private PlayerEntity player;
 
     @Override
     public void populateComponentRecipeFinder(ComponentRecipeMatcher finder) {
-        ((ComponentRecipeInputProvider)this.craftingInput).provideComponentRecipeInputs(finder);
+        ((ComponentRecipeInputProvider)this.input).provideComponentRecipeInputs(finder);
     }
 
     @Override
     public boolean matchesWithComponents(RecipeEntry<? extends Recipe<RecipeInputInventory>> recipe) {
-        return recipe.value().matches(this.craftingInput, this.owner.getWorld());
+        return recipe.value().matches(this.input, this.player.getWorld());
     }
 }
