+++ src/main/java/net/crioch/fifymcc/mixin/screen/PlayerScreenHandlerMixin.java
@@ -1,4 +1,4 @@
-package net.crioch.fifymcc.mixin.screenhandler;
+package net.crioch.fifymcc.mixin.screen;
 
 import net.crioch.fifymcc.recipe.ComponentRecipeMatcher;
 import net.crioch.fifymcc.interfaces.ComponentRecipeInputProvider;
@@ -7,28 +7,28 @@ import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.inventory.RecipeInputInventory;
 import net.minecraft.recipe.Recipe;
 import net.minecraft.recipe.RecipeEntry;
-import net.minecraft.screen.CraftingScreenHandler;
+import net.minecraft.screen.PlayerScreenHandler;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 
-@Mixin(CraftingScreenHandler.class)
-public class CraftingScreenHandlerMixin implements ComponentRecipeScreenHandler<RecipeInputInventory> {
+@Mixin(PlayerScreenHandler.class)
+public class PlayerScreenHandlerMixin implements ComponentRecipeScreenHandler<RecipeInputInventory> {
     @Final
     @Shadow
-    private RecipeInputInventory input;
+    private RecipeInputInventory craftingInput;
 
     @Final
     @Shadow
-    private PlayerEntity player;
+    private PlayerEntity owner;
 
     @Override
     public void populateComponentRecipeFinder(ComponentRecipeMatcher finder) {
-        ((ComponentRecipeInputProvider)this.input).provideComponentRecipeInputs(finder);
+        ((ComponentRecipeInputProvider)this.craftingInput).provideComponentRecipeInputs(finder);
     }
 
     @Override
     public boolean matchesWithComponents(RecipeEntry<? extends Recipe<RecipeInputInventory>> recipe) {
-        return recipe.value().matches(this.input, this.player.getWorld());
+        return recipe.value().matches(this.craftingInput, this.owner.getWorld());
     }
 }
